+++ lib/hadupils.rb
@@ -3,9 +3,11 @@ module Hadupils
 end
 
 require 'hadupils/assets'
-require 'hadupils/commands'
+require 'hadupils/helpers'
 require 'hadupils/extensions'
 require 'hadupils/runners'
 require 'hadupils/search'
 require 'hadupils/util'
 require 'hadupils/hacks'
+
+require 'hadupils/commands'